feat(Next): Move Gallery List to next; Unit test needs next server running (manually)
