+++ api/server/plugins/editAlbum/lib/index.js
@@ -1,4 +1,4 @@
-const gallery = require('../../gallery/lib/gallery');
+const gallery = require('../../../../../app/pages/api/galleryList/gallery');
 const routes = require('../../../lib/routes');
 const validation = require('../../../lib/validation');
