+++ app/__tests__/api/galleryList/gallery.js
@@ -1,0 +1,9 @@
+const lib = require('../../../pages/api/galleryList/gallery');
+
+describe('Gallery Library', () => {
+  test('* Get Galleries', async () => {
+    const galleries = await lib.getGalleries();
+    expect(galleries.length).toBeGreaterThan(0);
+    expect(galleries.includes('demo')).toBeTruthy();
+  });
+});
