+++ app/package.json
@@ -20,6 +20,7 @@
     "babel-jest": "^25.1.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^25.1.0",
+    "node-fetch": "^2.6.1",
     "react-test-renderer": "^16.12.0"
   },
   "license": "MIT"
