+++ app/pages/api/galleryList/gallery.js
@@ -1,7 +1,5 @@
 const fs = require('fs');
 
-const utils = require('../../utils');
-
 const fsPromises = fs.promises;
 
 async function getGalleries() {
@@ -9,8 +7,7 @@ async function getGalleries() {
     const hasPrefix = (content) => content.isDirectory();
     const namesOnly = (content) => content.name;
 
-    const galleriesPath = utils.file.safePublicPath('/galleries');
-    const contents = await fsPromises.readdir(galleriesPath, { withFileTypes: true });
+    const contents = await fsPromises.readdir('../public/galleries', { withFileTypes: true });
 
     const galleries = contents.filter(hasPrefix).map(namesOnly);
